Synchronize master data

+++ src/core/RefractionExtinction.hpp
@@ -60,8 +60,8 @@ public:
 	//! Note that forward/backward are no absolute reverse operations!
 	void forward(const Vec3d& altAzPos, float* mag) const
 	{
-		Q_ASSERT(std::fabs(altAzPos.length()-1.f)<0.001f);
-		*mag += airmass(altAzPos[2], false) * ext_coeff;
+		Q_ASSERT(std::fabs(altAzPos.length()-1.)<0.001);
+		*mag += airmass(static_cast<float>(altAzPos[2]), false) * ext_coeff;
 	}
 	
 	void forward(const Vec3f& altAzPos, float* mag) const
@@ -76,7 +76,7 @@ public:
 	//! Note that forward/backward are no absolute reverse operations!
 	void backward(const Vec3d& altAzPos, float* mag) const
 	{
-		*mag -= airmass(altAzPos[2], false) * ext_coeff;
+		*mag -= airmass(static_cast<float>(altAzPos[2]), false) * ext_coeff;
 	}
 	
 	void backward(const Vec3f& altAzPos, float* mag) const
